requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydub>=0.25.1
playsound>=1.3.0
python-docx>=1.0.0
ffmpeg-python>=0.2.0 